+++ lib/pleroma/web/activity_pub/activity_pub.ex
@@ -1077,7 +1077,7 @@ defmodule Pleroma.Web.ActivityPub.ActivityPub do
           data
         end
 
-      Repo.insert(%Object{data: obj_data})
+      Object.insert(%Object{data: obj_data})
     end
   end