+++ priv/repo/migrations/20191007123056_add_ap_id_column_to_objects.exs
@@ -1,0 +1,9 @@
+defmodule Pleroma.Repo.Migrations.AddApIdColumnToObjects do
+  use Ecto.Migration
+
+  def change do
+    alter table(:objects) do
+      add :ap_id, :string
+    end
+  end
+end